+++ src/mongo_module/modules/learner_progress/learner_progress.router.ts
@@ -1,12 +1,13 @@
 import { Router } from "express";
 import verifyToken from '../../middlewares/verify.token';
 import learnerProgressController from "./learner_progress.controller";
+import versionedAuth from "../../middlewares/versionedAuth";
 
 
 const lessonRouter = Router();
 
-lessonRouter.post("/createLearnerProgress", verifyToken, learnerProgressController.createLearnerProgress);
+lessonRouter.post("/createLearnerProgress", versionedAuth, learnerProgressController.createLearnerProgress);
 
-lessonRouter.get("/learnerProgressByuserId", verifyToken, learnerProgressController.learnerProgressByuserId);
+lessonRouter.get("/learnerProgressByuserId/:userId?", versionedAuth, learnerProgressController.learnerProgressByuserId);
 
 export default lessonRouter;
