+++ src/mongo_module/modules/mozhigal_tracker/mozhigal_tracker.router.ts
@@ -1,14 +1,14 @@
 import { Router } from "express";
-import verifyToken from '../../middlewares/verify.token';
 import MozhigalTrackerController from "./mozhigal_tracker.controller";
+import versionedAuth from "../../middlewares/versionedAuth";
 
 const mozhigalTrackerRouter = Router();
 
-mozhigalTrackerRouter.post("/:lessonId", verifyToken, MozhigalTrackerController.addLearningLogs);
+mozhigalTrackerRouter.post("/:lessonId", versionedAuth, MozhigalTrackerController.addLearningLogs);
 
-mozhigalTrackerRouter.get("/student", verifyToken, MozhigalTrackerController.getCumulativeScore);
+mozhigalTrackerRouter.get("/student/:userId?", versionedAuth, MozhigalTrackerController.getCumulativeScore);
 
-mozhigalTrackerRouter.get("/lessons", verifyToken, MozhigalTrackerController.getLessonWiseScore);
+mozhigalTrackerRouter.get("/lessons/:userId?", versionedAuth, MozhigalTrackerController.getLessonWiseScore);
 
 
 export default mozhigalTrackerRouter;
